feat: Fixed some bugs

+++ backend/src/services/todo.service.ts
@@ -1,28 +1,43 @@
+import SubTask from '../models/SubTask';
 import Todo from '../models/Todo';
 import { ITodo } from '../types/todos.type';
 
 export class TodoService {
-  static async create(data: ITodo) {
-    return Todo.build(data);
+  static async create(userId: string, data: ITodo) {
+    return await Todo.build({ ...data, userId }).save();
   }
 
-  static async update(id: string, data: ITodo) {
-    return Todo.update(data, { where: { id } });
+  static async update(userId: string, id: string, data: ITodo) {
+    return await Todo.update(data, { where: { id, userId } });
   }
 
-  static async delete(id: string) {
-    return Todo.destroy({ where: { id } });
+  static async delete(userId: string, id: string) {
+    return await Todo.destroy({ where: { id, userId } });
   }
 
-  static async findAll() {
-    return Todo.findAll();
+  static async getAll(userId: string) {
+    return await Todo.findAll({
+      where: { userId },
+      include: [
+        {
+          model: SubTask,
+        },
+      ],
+    });
   }
 
-  static async getAllTodosByUser(userId: string): Promise<ITodo[] | null> {
-    return await Todo.findAll({ where: { userId } });
-  }
-
-  static async findById(id: string) {
-    return Todo.findOne({ where: { id } });
+  static async getById(userId: string, id: string): Promise<ITodo | null> {
+    return await Todo.findOne({
+      where: { id, userId },
+      include: [
+        {
+          model: SubTask,
+          where: {
+            todoId: id,
+            userId,
+          },
+        },
+      ],
+    });
   }
 }
